+++ code/packages/rust/html-parser/tests/browser_content_tree_test.rs
@@ -95,6 +95,8 @@ struct ExpectedContentNode {
     #[serde(default)]
     image_map_coords: Option<String>,
     #[serde(default)]
+    usemap: Option<String>,
+    #[serde(default)]
     srcset: Option<String>,
     #[serde(default)]
     resolved_srcset: Option<String>,
@@ -457,6 +459,7 @@ impl ExpectedContentNode {
             image_map_name: self.image_map_name,
             image_map_shape: self.image_map_shape,
             image_map_coords: self.image_map_coords,
+            usemap: self.usemap,
             srcset: self.srcset,
             resolved_srcset: self.resolved_srcset,
             sizes: self.sizes,